requirements.txt
flask
flask-cors
gunicorn
pillow
torch
torchvision
gdown
pandas